Bugfix: ObtainCert cannot look at the Managed flag, so we should do that instead
While using caddy-net, I discovered that none of the TLS directives seemed to work (tls off, tls self_signed, ...). However, setting the email address did work. The problem is that a certificate is fetched regardless of whether the TLS is being managed or not. This is incorrect and will not work on development setups where ACME fails.